requirements.txt
langchain
langchain-openai
openai
pypdf
chromadb
tiktoken
pysqlite3
streamlit-extras